move packagable code to utils
