+++ lib/create-webpack-config.js
@@ -8,7 +8,7 @@ const resolvePkg = require('resolve-pkg');
 const urlJoin = require('url-join');
 const fs = require('fs');
 
-const safeWebpackUglifier = require('./packageable/safe-webpack-uglifier');
+const safeWebpackUglifier = require('./utils/safe-webpack-uglifier');
 const { WEBPACK_ASSETS_BASENAME } = require('./constants');
 
 function createWebpackConfig(urc) {
