+++ lib/public-files-copier.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const autoCopy = require('./packageable/auto-copy');
+const autoCopy = require('./utils/auto-copy');
 
 module.exports = copyPublicFiles;
