+++ lib/start-server.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const reloadingServer = require('./packageable/reloading-server');
+const reloadingServer = require('./utils/reloading-server');
 const logger = require('./logger');
 
 function startServer(urc) {
