+++ lib/utils/auto-copy.js
