+++ lib/utils/safe-webpack-uglifier.js
